- Modified Create tournament form to have some padding and correct button colors. - Moved create tournament to component.
